Add files via upload
Interactive map: For this data vis proposal an interactive map would be created to show all of the lowest income, rural towns in America. Upon clicking on a map marker, one would be able to view the median income, town population, and nearest distance to the grocery on the map. A more advanced map would have a feature that allows you to type in your information (zip code or median income) to see how your community measures up against the lowest income, rural towns. ZeeMaps was used to create this map. 

Bar graph: The bar graph would show the average travel length to the nearest grocery store for low income, rural residents across 50 states. Tableau was used to create this map.